requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-dotenv
apscheduler
watchdog
croniter